src/embedded/requirements.txt
pydobot==1.3.2
pyserial==3.4
